+++ meraki_client/aio/__init__.py
@@ -43,7 +43,7 @@ __all__ = [
     "BaseURL",
     "types",
 ]
-__version__ = "0.15.1"
+__version__ = "0.15.2"
 __api_version__ = "v1.70.0"